requirements.txt
mysql-connector-python
pyyaml
paramiko